pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "klipper-console"
version = "0.1.0"
description = "Terminal-native CLI for Klipper via Moonraker"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
  {name = "Carlos Perez", email = "carlos@darkoperator.com"}
]
keywords = [
  "klipper",
  "moonraker",
  "3d-printing",
  "cli",
  "terminal",
  "repl",
  "automation"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Home Automation",
  "Topic :: System :: Hardware",
  "Topic :: Utilities"
]

dependencies = [
  "prompt_toolkit>=3.0",
  "httpx>=0.27",
  "rich>=13.7",
  "websocket-client>=1.6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "black>=23.0",
  "mypy>=1.0",
  "ruff>=0.1.0"
]

[project.urls]
Homepage = "https://github.com/darkoperator/Klipper-Console"
Documentation = "https://github.com/darkoperator/Klipper-Console/blob/main/README.md"
Repository = "https://github.com/darkoperator/Klipper-Console"
Issues = "https://github.com/darkoperator/Klipper-Console/issues"
Changelog = "https://github.com/darkoperator/Klipper-Console/blob/main/CHANGELOG.md"

[project.scripts]
klipper-console = "klipper_console.cli:main"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
